third_party/WebKit/LayoutTests/accessibility/aria-disabled.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body id="body">
<input id="text1" type="text" aria-disabled="true" size=20>
<input id="text2" type="text" aria-disabled="false" size=20>
<div role="group" id="group1" aria-disabled="true">
  <div role="group">
    <a id="link1" href="#">link</a>
    <button id="button1" type="submit" aria-disabled="false">submit</button>
  <div>
<div>
</div>
<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that the aria-disabled attribute works. The text field should not be enabled.");

    if (window.accessibilityController) {

          var textField = accessibilityController.accessibleElementById("text1");
          shouldBe("textField.isEnabled", "false");

          textField = accessibilityController.accessibleElementById("text2");
          shouldBe("textField.isEnabled", "true");

          var group = accessibilityController.accessibleElementById("group1");
          shouldBe("group.isEnabled", "false");

          var link = accessibilityController.accessibleElementById("link1");
          shouldBe("link.isEnabled", "false");

          var button = accessibilityController.accessibleElementById("button1");
          shouldBe("button.isEnabled", "true");
    }

</script>

</body>
</html>
